Created a package HandlingPOPUPs and moved all the testcases related to it.
