+++ Desktop/eclipse/SeleniumProject/SampleProgram/src/BasicJavaPrograms/HandlingPOPUPs/Handling_ConfirmationPopUp.java
@@ -1,4 +1,4 @@
-package BasicJavaPrograms;
+package BasicJavaPrograms.HandlingPOPUPs;
 
 import java.util.concurrent.TimeUnit;
