--- Desktop/eclipse/SeleniumProject/SampleProgram/src/BasicJavaPrograms/Handling_PromptMessage.java
@@ -1,9 +1,0 @@
-package BasicJavaPrograms;
-
-public class Handling_PromptMessage {
-
-	public static void main(String[] args) {
-
-	}
-
-}
